test: Rename control-plane to controlplane
Fix the naming so that the package does not need to be imported with
a specific name.
