+++ test/controlplane/services/dual-stack/generate.sh
