+++ test/controlplane/services/dual-stack/manifests/kind-config-1.24.yaml
