+++ test/controlplane/services/dual-stack/run.go
@@ -9,8 +9,8 @@ import (
 	fakeDatapath "github.com/cilium/cilium/pkg/datapath/fake"
 	lb "github.com/cilium/cilium/pkg/loadbalancer"
 	"github.com/cilium/cilium/pkg/option"
-	controlplane "github.com/cilium/cilium/test/control-plane"
-	. "github.com/cilium/cilium/test/control-plane/services"
+	"github.com/cilium/cilium/test/controlplane"
+	. "github.com/cilium/cilium/test/controlplane/services"
 )
 
 func RunDualStackTestWithVersion(t *testing.T, version string) {
