+++ test/controlplane/services/dual-stack/v1.20/dualstack_test.go
@@ -6,7 +6,7 @@ package v1_20
 import (
 	"testing"
 
-	dualstack "github.com/cilium/cilium/test/control-plane/services/dual-stack"
+	dualstack "github.com/cilium/cilium/test/controlplane/services/dual-stack"
 )
 
 func TestDualStack1_20(t *testing.T) {
