+++ test/controlplane/services/dual-stack/v1.20/lbmap1.golden
