+++ test/controlplane/services/dual-stack/v1.22/init.yaml
