+++ test/controlplane/services/dual-stack/v1.22/lbmap1.golden
