+++ test/controlplane/services/dual-stack/v1.24/init.yaml
