+++ test/controlplane/services/dual-stack/v1.24/lbmap1.golden
