+++ test/controlplane/services/graceful-termination/generate.sh
