+++ test/controlplane/services/graceful-termination/lbmap1.golden
