+++ test/controlplane/services/graceful-termination/lbmap2.golden
