+++ test/controlplane/services/graceful-termination/lbmap3.golden
