+++ test/controlplane/services/nodeport/v1.20/lbmap1.golden
