+++ test/controlplane/services/nodeport/v1.22/lbmap1.golden
