+++ test/controlplane/services/nodeport/v1.24/lbmap1.golden
